# openapi.yaml
openapi: 3.0.3
info:
  title: ShopFloor OS
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8787
paths:
  /health:
    get:
      security: []
      responses:
        "200":
          description: process up
  /v1/commands/run.start:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetId, workOrderId, operationId]
              properties:
                assetId: { type: string }
                workOrderId: { type: string }
                operationId: { type: string }
      responses:
        "200":
          description: event appended or idempotent replay
  /v1/commands/qty.scrap:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetId, workOrderId, operationId, qty, reasonCode]
              properties:
                assetId: { type: string }
                workOrderId: { type: string }
                operationId: { type: string }
                qty: { type: integer, minimum: 1 }
                reasonCode: { type: string }
      responses:
        "200":
          description: scrap recorded
  /v1/commands/downtime.start:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetId, reasonCode]
              properties:
                assetId: { type: string }
                reasonCode: { type: string }
      responses:
        "200":
          description: downtime opened
  /v1/commands/run.complete:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetId]
              properties:
                assetId: { type: string }
      responses:
        "200":
          description: run closed, lock released
  /v1/commands/run.pause:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetId]
              properties:
                assetId: { type: string }
      responses:
        "200":
          description: run paused; lock remains
  /v1/commands/run.resume:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetId]
              properties:
                assetId: { type: string }
      responses:
        "200":
          description: run resumed
  /v1/commands/qty.good:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetId, workOrderId, operationId, qty]
              properties:
                assetId: { type: string }
                workOrderId: { type: string }
                operationId: { type: string }
                qty: { type: integer, minimum: 1 }
      responses:
        "200":
          description: good qty recorded
  /v1/commands/downtime.end:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetId]
              properties:
                assetId: { type: string }
      responses:
        "200":
          description: downtime closed
  /v1/commands/handoff.submit:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: handoff offered
  /v1/commands/handoff.accept:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: handoff accepted (supervisor/planner)
  /v1/commands/handoff.override:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fromShift, toShift, reason]
              properties:
                fromShift: { type: string }
                toShift: { type: string }
                reason: { type: string }
      responses:
        "200":
          description: pending handoff cleared without accept
  /v1/catalog/assets:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name]
              properties:
                id: { type: string }
                code: { type: string }
                name: { type: string }
      responses:
        "200":
          description: asset created (planner)
  /v1/catalog/reason-codes:
    post:
      responses:
        "200":
          description: reason code created (planner)
  /v1/catalog/work-orders:
    post:
      responses:
        "200":
          description: work order created; emits work_order.opened
  /v1/catalog/operations:
    post:
      responses:
        "200":
          description: routing step created (planner)
  /v1/commands/record.correct:
    post:
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [replacesEventId, reason]
              properties:
                replacesEventId: { type: string }
                reason: { type: string }
      responses:
        "200":
          description: correction appended; original row unchanged
  /v1/tape:
    get:
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: plant event tape (auditor bearer only)
        "403":
          description: non-auditor
  /v1/floor:
    get:
      responses:
        "200":
          description: plant board payload
  /v1/plants/{plantId}/assets/{assetId}/live:
    get:
      parameters:
        - name: plantId
          in: path
          required: true
          schema: { type: string }
        - name: assetId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: open run and downtime
  /v1/work-orders/{id}/timeline:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: event stream for a work order
  /v1/metrics/oee:
    get:
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: asset
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OEE-lite for one asset, or assets[] when asset is omitted. Null factors are omitted math, not zero.
        "400":
          description: missing or inverted window
        "404":
          description: unknown asset
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security:
  - bearer: []